+++ src/main/java/com/ev/scm/service/impl/FundInitializationServiceImpl.java
@@ -27,7 +27,7 @@ public class FundInitializationServiceImpl implements FundInitializationService 
 
 
 	@Override
-	public FundInitializationDO get(Integer id){
+	public FundInitializationDO get(Long id){
 		return fundInitializationDao.get(id);
 	}
